requirements.txt
langchain
langchain-google-genai
python-dotenv
google-generativeai
langchain-community
gradio
opik